feat: search,view product

+++ thinkpro/src/main/java/com/gr25/thinkpro/repository/CartRepository.java
@@ -8,8 +8,9 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartRepository extends JpaRepository<Cart, Integer> {
+public interface CartRepository extends JpaRepository<Cart, Long> {
     @Modifying
     @Query("DELETE FROM Cart c WHERE c.customer.customerId = :customerId")
     void deleteByCustomerId(@Param("customerId") long customerId);
+
 }
